pyproject.toml
[project]
name = "aisearch"
version = "0.1.0"
description = "AI Search Project"
authors = [{ name = "emias" }]
requires-python = ">=3.12,<4.0"
dependencies = []

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["logs*", "volumes*", "info*"]

[project.optional-dependencies]
api = [
    "dishka>=1.6.0,<2",
    "pydantic>=2.12.4,<3.0.0",
    "structlog>=25.4.0,<26",
    "redis>=6.2.0,<7",
    "pydantic-settings>=2.11.0,<3.0.0",
    "fastapi[all]>=0.115.12,<0.116",
    "slowapi>=0.1.9,<0.2",
    "toml>=0.10.2,<0.11",
    "uvicorn-worker>=0.4.0,<0.5",
    "celery>=5.5.3,<6",
    "asyncpg>=0.30.0,<0.31.0",
    "sqlalchemy>=2.0.41,<3.0.0",
    "alembic>=1.17.1,<2.0.0",
    "sqlalchemy-utils>=0.42.0,<0.43.0",
    "psycopg2-binary>=2.9.11,<3.0.0",
    "alembic-postgresql-enum>=1.8.0,<2.0.0",
    "httpx>=0.28.1,<0.29.0",
]

queue = [
    "dishka>=1.6.0,<2",
    "pydantic>=2.11.7,<3.0.0",
    "structlog>=25.4.0,<26",
    "redis>=6.2.0,<7",
    "celery>=5.5.3,<6",
    "pydantic-settings>=2.10.1,<3.0.0",
]

search = [
    "dishka>=1.6.0,<2",
    "structlog>=25.4.0,<26",
    "toml>=0.10.2,<0.11",
    "uvicorn-worker>=0.4.0,<0.5",
    "sqlalchemy>=2.0.41,<3.0.0",
    "pymilvus>=2.5.10,<3",
    "fastapi[all]>=0.115.12,<0.116",
    "sentence-transformers==5.1.0",
    "redis>=6.2.0,<7",
    "nltk>=3.9.1,<4",
    "pymorphy3>=2.0.4,<3",
    "celery>=5.5.3,<6",
    "aiohttp>=3.12.14,<4",
    "openpyxl>=3.1.5,<4",
    "pandas>=2.3.1,<3",
    "shortuuid>=1.0.13,<2",
    "aiofiles>=24.1.0,<25",
    "opensearch>=0.9.2,<0.10",
    "opensearch-py>=3.0.0,<4",
    "torch==2.7.1",
    "torchvision==0.22.1",
    "torchaudio==2.7.1",
    "pyarrow>=21.0.0,<22",
    "slowapi>=0.1.9,<0.2",
    "psutil>=7.1.1,<8.0.0",
    "apscheduler>=3.1.5",
    "alembic>=1.17.1,<2.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "sqlalchemy-utils>=0.42.0,<0.43.0",
    "psycopg2-binary>=2.9.11,<3.0.0",
    "pydantic>=2.12.4,<3.0.0",
    "pydantic-settings>=2.11.0,<3.0.0",
    "httpx>=0.28.1,<0.29.0",
    "alembic-postgresql-enum>=1.8.0,<2.0.0",
    "python-dateutil>=2.9.0.post0,<3.0.0",
]

dev = [
    "black>=24.10.0,<25",
    "debugpy>=1.8.12,<2",
    "pytest>=8.3.4,<9",
    "ruff>=0.6.6,<0.7",
    "pytest-asyncio>=0.21.1,<0.22",
    "pytest-cov>=5.0.0,<6",
    "pytest-xdist>=3.6.1,<4",
    "commitizen>=3.27.0,<4",
    "pytest-dotenv>=0.5.2,<0.6",
    "factory-boy>=3.3.0,<4",
    "faker>=26.0.0,<27",
    "pytest-lazy-fixture>=0.6.3,<0.7",
    "pytest-mock>=3.14.0,<4",
    "asgi-lifespan>=2.1.0,<3",
    "anyio>=4.8.0,<5",
    "polyfactory>=2.19.0,<3",
    "mypy>=1.15.0,<2",
    "types-toml>=0.10.8.20240310,<0.11",
    "testcontainers>=4.10.0,<5",
    "pre-commit>=4.2.0,<5",
]
